"Inico de implementacion de mercado pago"

+++ E-commerce_Project/src/main/java/com/ecommerce/ecommerce/dto/MercadoPagoItemDTO.java
@@ -1,6 +1,8 @@
 package com.ecommerce.ecommerce.dto;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MercadoPagoItemDTO {
     private String id;
@@ -9,5 +11,5 @@ public class MercadoPagoItemDTO {
     private String pictureUrl;
     private String categoryId;
     private Integer quantity;
-    private Double unitPrice;
+    private BigDecimal unitPrice; // <-- Debe ser BigDecimal en Java
 }
